Ajouts vérif sur droits services

+++ src/Controller/ServicesController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Services;
 use App\Form\ServicesType;
+use App\Service\User as UserService;
 use App\Repository\ServicesRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -14,9 +15,19 @@ use Symfony\Component\Routing\Attribute\Route;
 #[Route('/services')]
 final class ServicesController extends AbstractController
 {
+
+    public function __construct(private UserService $userService)
+    {
+    }
+
     #[Route(name: 'app_services_index', methods: ['GET'])]
     public function index(ServicesRepository $servicesRepository): Response
     {
+        if($this->userService->isAllowedToGo($this->getUser(),"SERVICE_CONSULT") == false)
+        {
+            return $this->redirectToRoute('app_index');
+        }
+
         return $this->render('services/index.html.twig', [
             'services' => $servicesRepository->findAll(),
         ]);
@@ -25,6 +36,11 @@ final class ServicesController extends AbstractController
     #[Route('/new', name: 'app_services_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if($this->userService->isAllowedToGo($this->getUser(),"SERVICE_CREATE") == false)
+        {
+            return $this->redirectToRoute('app_index');
+        }
+
         $service = new Services();
         $form = $this->createForm(ServicesType::class, $service);
         $form->handleRequest($request);
@@ -45,6 +61,11 @@ final class ServicesController extends AbstractController
     #[Route('/{id}', name: 'app_services_show', methods: ['GET'])]
     public function show(Services $service): Response
     {
+        if($this->userService->isAllowedToGo($this->getUser(),"SERVICE_CONSULT") == false)
+        {
+            return $this->redirectToRoute('app_index');
+        }
+
         return $this->render('services/show.html.twig', [
             'service' => $service,
         ]);
@@ -53,6 +74,11 @@ final class ServicesController extends AbstractController
     #[Route('/{id}/edit', name: 'app_services_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Services $service, EntityManagerInterface $entityManager): Response
     {
+        if($this->userService->isAllowedToGo($this->getUser(),"SERVICE_CREATE_UPDATE") == false)
+        {
+            return $this->redirectToRoute('app_index');
+        }
+
         $form = $this->createForm(ServicesType::class, $service);
         $form->handleRequest($request);
 
@@ -71,6 +97,11 @@ final class ServicesController extends AbstractController
     #[Route('/{id}', name: 'app_services_delete', methods: ['POST'])]
     public function delete(Request $request, Services $service, EntityManagerInterface $entityManager): Response
     {
+        if($this->userService->isAllowedToGo($this->getUser(),"SERVICE_DELETE") == false)
+        {
+            return $this->redirectToRoute('app_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$service->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($service);
             $entityManager->flush();
